requirements.txt
aiohttp==3.9.5
discord.py==2.3.2
python-dotenv==1.1.0
piexif==1.1.3
pillow==10.3.0
fastapi==0.128.5
pydantic==2.11.3
uvicorn==0.40.0
slowapi==0.1.9
